Added offsets and transformationset
git-svn-id: https://svn.mech.kuleuven.be/repos/orocos/orocos-trunk@371 ce417995-dfc9-0310-95a0-acdaff106893

+++ packages/device_drivers/jr3/current/src/JR3DSP.h
@@ -86,5 +86,5 @@ void JR3DSP_getDataFromFilter6(struct ForceArray* data, unsigned int dsp);
 // ------------------------------------
 // A rather specific tranformation: rotation and translation around the
 // z-axix; (deg, m)
-void JR3DSP_transformCoordinateSystem(double angle, double distance, unsigned int dsp);
+void JR3DSP_transformCoordinateSystem(float angle, float distance, unsigned int dsp);
 #endif // JR3DSP_H
